+++ test/foundry/TestHelpers.sol
@@ -103,7 +103,7 @@ abstract contract TestHelpers is AssertionHelpers, TestParameters {
             maximumEntries: 200,
             maximumEntriesPerParticipant: 200,
             prizesTotalValue: 1 ether,
-            minimumProfitBp: uint16(500),
+            minimumProfitBp: 500,
             feeTokenAddress: address(0),
             prizes: prizes,
             pricingOptions: pricingOptions
